GYM-241: Get exercise from db to have its id in test data
