+++ e2e/test-data/exercises.data.ts
@@ -16,59 +16,3 @@ export type RoutineExercise = {
   exercise: Exercise;
   orderNumber: number;
 };
-
-export const noExercises: RoutineExercise[] = [];
-
-export const oneExercise: RoutineExercise[] = [
-  {
-    exerciseId: '383ec35d-a9f5-4c0d-84f7-c7f9344d0092',
-    defaultSets: 3,
-    defaultReps: 12,
-    defaultWeight: 40,
-    defaultRestTime: 90,
-    notes: '',
-    restTimeUnit: 'seconds',
-    weightUnit: 'kg',
-    routineExerciseId: '383ec35d-a9f5-4c0d-84f7-c7f9344d0092',
-    exercise: {
-      id: '383ec35d-a9f5-4c0d-84f7-c7f9344d0092',
-      name: 'Sit Squats',
-    },
-    orderNumber: 1,
-  },
-];
-
-export const twoExercises: RoutineExercise[] = [
-  {
-    exerciseId: '383ec35d-a9f5-4c0d-84f7-c7f9344d0092',
-    defaultSets: 3,
-    defaultReps: 12,
-    defaultWeight: 40,
-    defaultRestTime: 90,
-    notes: '',
-    restTimeUnit: 'seconds',
-    weightUnit: 'kg',
-    routineExerciseId: '383ec35d-a9f5-4c0d-84f7-c7f9344d0092',
-    exercise: {
-      id: '383ec35d-a9f5-4c0d-84f7-c7f9344d0092',
-      name: 'Sit Squats',
-    },
-    orderNumber: 1,
-  },
-  {
-    exerciseId: '718539a9-5054-43a3-9176-990fdf7a193b',
-    defaultSets: 3,
-    defaultReps: 12,
-    defaultWeight: 40,
-    defaultRestTime: 90,
-    notes: '',
-    restTimeUnit: 'seconds',
-    weightUnit: 'kg',
-    routineExerciseId: '718539a9-5054-43a3-9176-990fdf7a193b',
-    exercise: {
-      id: '718539a9-5054-43a3-9176-990fdf7a193b',
-      name: 'Single-Cone Sprint Drill',
-    },
-    orderNumber: 2,
-  },
-];
